修复 solon-boot-smarthttp 适配，在 contentLength(long) 时会精度丢失的问题

+++ solon-projects/solon-boot/solon-boot-smarthttp/src/main/java/org/noear/solon/boot/smarthttp/http/SmHttpContext.java
@@ -379,7 +379,7 @@ public class SmHttpContext extends WebContextBase {
 
     @Override
     public void contentLength(long size) {
-        _response.setContentLength((int) size);
+        _response.setContentLength(size);
     }
 
     @Override
